requirements.txt
Django>=4.2
channels>=4.0
daphne>=4.0
channels-redis>=4.0
djangorestframework-simplejwt>=5.3.0
redis>=5.0.1
python-dotenv>=1.1.1